requirements.txt
aiocron==1.3
pylint==2.6.0
python-dotenv==0.15.0
requests==2.25.0
Telethon==1.17.5
urllib3==1.26.2
